+++ src/network/CNeuron.java
@@ -1,7 +1,7 @@
 package network;
 
 public class CNeuron extends CBaseNeuron {
-	public void calculate(Double val){
+	public void calculate(float val){
 		super.setValue(val);
 	}
 }
