+++ src/network/neuron/CLinearNeuron.java
@@ -3,23 +3,23 @@ package network.neuron;
 import network.CBaseNeuron;
 
 public class CLinearNeuron extends CBaseNeuron{
-	private Double gradient;
-	private Double intersection;
+	private float gradient;
+	private float intersection;
 	
 	public CLinearNeuron(){
-		gradient = 1.0;
-		intersection = 0.0;
+		gradient = 1.0f;
+		intersection = 0.0f;
 	}
 	
-	public void calculate(Double val) {
+	public void calculate(float val) {
 		super.setValue(gradient * val + intersection);
 	}
 	
-	public void setGradient(Double g){
+	public void setGradient(float g){
 		gradient = g;
 	}
 	
-	public void setIntersection(Double i){
+	public void setIntersection(float i){
 		intersection = i;
 	}
 }
