+++ src/network/neuron/CLogisticNeuron.java
@@ -4,8 +4,8 @@ import network.CBaseNeuron;
 
 public class CLogisticNeuron extends CBaseNeuron {
 
-	public void calculate(Double val) {
-		super.setValue(1.0/(1.0+Math.pow(Math.E, -val)));
+	public void calculate(float val) {
+		super.setValue(1.0f/(1.0f+(float)Math.pow(Math.E, -val)));
 	}
 	
 }
