+++ src/network/neuron/CUnsignedLinearNeuron.java
@@ -1,11 +1,11 @@
 package network.neuron;
 
 public class CUnsignedLinearNeuron extends CLinearNeuron {
-	public void calculate(Double val) {
+	public void calculate(float val) {
 		if(val >= 0){
 			super.calculate(val);
 		}else{
-			super.setValue(0.0);
+			super.setValue(0.0f);
 		}
 	}
 }
